+++ resources/views/layouts/auth/card.blade.php
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
-    <x-site.head :title="isset($title) ? $title : null" />
+    <x-site.head :title="isset($title) ? $title : null" :site-custom-css="false" />
 
     <body class="min-h-screen bg-neutral-100 antialiased dark:bg-linear-to-b dark:from-neutral-950 dark:to-neutral-900">
         <div class="bg-muted flex min-h-svh flex-col items-center justify-center gap-6 p-6 md:p-10">